final-project/backend/requirements.txt
flask
flask-cors
transformers
diffusers
torch
pillow
moviepy
gtts
fpdf
